+++ tests/AdagucTests/TestGenericImageWarper.py
@@ -118,7 +118,7 @@ class TestGenericImageWarper(unittest.TestCase):
                 AdagucTestTools().compareImage(
                     self.expectedoutputsspath + filename,
                     self.testresultspath + filename,
-                    30,
+                    50,
                 )
             )