+++ adagucserverEC/CLegendRenderers/CCreateLegend.cpp
@@ -20,6 +20,11 @@ int CCreateLegend::createLegend(CDataSource *dataSource, CDrawImage *legendImage
     CStyleConfiguration *styleConfiguration = dataSource->getStyle();
     if (styleConfiguration != NULL) {
       if (styleConfiguration->styleConfig != NULL) {
+        auto drawSettings = getDrawFunctionSettings(dataSource, legendImage, styleConfiguration);
+        if (drawSettings.drawgrid == false) {
+          legendImage->crop(4);
+          return 0;
+        }
         if (styleConfiguration->styleConfig->LegendGraphic.size() == 1) {
           if (styleConfiguration->styleConfig->LegendGraphic[0]->attr.value.empty() == false) {
             const char *fileName = styleConfiguration->styleConfig->LegendGraphic[0]->attr.value.c_str();
@@ -72,7 +77,7 @@ int CCreateLegend::createLegend(CDataSource *dataSource, CDrawImage *legendImage
     legendImage->crop(4);
     return 0;
   }
-
+  // If no min or mas is set, detect it. Also detect it if the legend has no fixed/min/max
   if (styleConfiguration->legendScale == 0.0f || styleConfiguration->legendHasFixedMinMax == false) {
     estimateMinMax = true;
   } else {
@@ -132,6 +137,10 @@ int CCreateLegend::createLegend(CDataSource *dataSource, CDrawImage *legendImage
     } else {
       legendType = discrete;
     }
+  }
+
+  if (styleConfiguration->renderMethod == RM_GENERIC && styleConfiguration->shadeIntervals.size() > 0) {
+    legendType = discrete;
   }
 
   if (styleConfiguration->featureIntervals.size() > 0) {