+++ NEWS.md
@@ -1,3 +1,11 @@
+**Version 6.0.0 - 2025-11-18**
+
+- Implemented generic image warper, available via RenderMethod 'generic'. This warper uses the already existing generic data warper and combines functionality of the nearest and bilinear renderer.
+- If no RenderMethod is given in the Style configuration, 'generic' is used by default
+- Styles can be included into each other via the `IncludeStyle` option: See example at [adaguc.tests.ahn_utrechtse_heuvelrug_500m.xml](./data/config/datasets/adaguc.tests.ahn_utrechtse_heuvelrug_500m.xml)
+- The WCS server now handles swapped x/y dimension axes properly.
+- Data postprocessor convert_uv_components (isVectorLike) is not added by default anymore.
+
 **Version 5.0.1 - 2025-11-05**
 
 ***Support to configure windbarbs and points with graphical stylings for specific values like windspeeds***